Allocate menus statically instead of on the heap
After doing some research, I've decided against using the heap as it can
cause problems on MCUs with a limited amount of memory. Aside from the
small overhead to manage the heap, frequent heap allocations under the
right conditions can lead to fragmentation of free heap space. Calls to
malloc will fail if there's no space big enough for the requested
allocation, which would be a hard to reproduce fatal problem.

Given that the intended use of this firmware is for timing fuel in an
aircraft, I'd like to avoid dynamic allocation completely if possible to
avoid potential hard to reproduce and debug failure modes.

Changing to static allocation also saves a decent chunk of flash - 678
bytes in the program and 14 bytes in data.