Define menu item types
This adds an enum for declaring and figuring out what each menu item can be
used for (mainly if it can display a value or not). These are a work in
progress...

This also renames progmem string variables to use the prefix "pstr_" instead
of "mnu_" (intended to be clearer and less specific to the menu). I'm planning
to reuse these progmem strings on a dedicated value setting screen, so have
renamed them.

These strings will likely move to a definitions file sooner or later.