Implement view stack and basic menu navigation
This adds an simple abstraction for managing a hierarchy of views.

I've moved menu items out of the stack for now as they need to be
used in multiple functions that are called independently. This ends up
using more SRAM than I'd like so I'm considering a few other options:

 - Have a global `menu_items items[largest_menu]` array to cap SRAM usage
   at the largest menu size. Currently each menu exclusively uses its own
   block of SRAM which will get pretty wasteful as the menu_item struct
   and the total number of menu items in the application grow.

 - Change the structure/interface of ViewStack -> ViewStackFrame to
   allow for definition of menu items on the stack that is accessible
   to all functions in a view. This might need to be combined with the
   first idea to get any SRAM or flash savings.