Store icons as structs in progmem
This updates the way icons are stored to use a single pointer, making the code
read a little clearer, easier to work with (since there's now a type for icons
instead of passing around loose parameters), and saves program space by storing
and passing around fewer values.

At the time of commit, this saved ~40 bytes of program space and a few bytes of
SRAM from dimensions now being stored in progmem.

I made this change as working with icon + size data was becoming annoying.