Remove typed editable flags: type is handled internally by implementations
This enum was created at a time I wanted to make editable values generic. I've
since decided that value editing is best implemented internally by each view in the
view's increment and decrement handlers. This allows handling limits, wrapping, and
special case behaviours without making a huge API to handle it all.

The editable flag now just signals to the menu that an item supports entering the
editing state, which the view can read to decide if inputs should edit or move the
menu's cursor.