Modify u8glib SSD1306 screen driver to only draw once per update
The SSD1306 driver that ships with u8glib calls draw 4 times to fill a 32 pixel
high page buffer. This minimises memory use, but it's been causing drawing to
be noticably laggy on some screens (eg. aircraft config).

SRAM usage is not currently a problem so I've made a version of this u8glib
driver that renders to a full-size buffer for the 128x64 screen. This is based
on the code in `u8g_dev_ssd1306_128x64.c` and `u8g_pb16v1.c`.

The display now uses 1KB of SRAM which could be reduced to 512 bytes with a
2-page buffer in future at the cost of some performance if extra SRAM is needed
for something.

The SPI mode for this driver does not currently work.