Add one-page driver for ST7565 monochrome LCD
This includes a clamped contrast value that fits with the contrast setting of
the display unit. I've been testing out this display for this project, but I'm
not sure if it'll be used or replaced with an OLED.

Need to trial both in real-world use to see if either is easier to see in
bright conditions.