Add crude initialisation status display during start-up
I've been having issues with flaky connections on a breadboard. This is so I
can see where the program is stuck after power-on if its blocked trying to
read from the F-RAM or RTC.

It would've been nice to have the screen buffer kept retained to reduce the
time taken drawing the whole status screen multiple times, but that would've
required some custom u8g messages and state which I didn't feel like editing
into the library.